% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{finalise_plot}
\alias{finalise_plot}
\title{Finalise Snare Industries standards plot for export}
\usage{
finalise_plot(plot_variable,
  bg_image_path = file.path(system.file("data", package = "snareplot"),
  "5k-botanical-minimal-potted-wallpaper.jpg"), draw_background)
}
\description{
Add Snare Industries Botanical background and Snare Industries Data Science watermark ready for image export
}
\examples{
df <- iris
g <- ggplot(data=df, aes(x=Petal.Length, y=Petal.Width))
g <- g + geom_point(aes(color=Species))
g <- g + theme_snare() + scale_colour_snare()
finalise_plot(g)
}
\keyword{finalise_plot}
