% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_snare.R
\name{scale_colour_snare}
\alias{scale_colour_snare}
\title{Add Snare Industries colours to ggplot chart}
\usage{
scale_colour_snare()
}
\description{
This function allows you to add the Snare Industries colour palette to your ggplotgraphics.
See also theme_snare() and scale_colour_snare
}
\examples{
df <- iris
g <- ggplot(data=df, aes(x=Petal.Length, y=Petal.Width))
g <- g + geom_point(aes(color=Species))
g +
theme_snare() +
scale_colour_snare()
}
\keyword{scale_colour_snare}
