% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_dcc.R
\name{scale_fill_dcc}
\alias{scale_fill_dcc}
\title{Add DCC colours to ggplot chart}
\usage{
scale_fill_dcc()
}
\description{
This function allows you to add the DCC colour palette to your ggplotgraphics.
See also theme_dcc() and scale_colour_dcc
}
\examples{
df <- iris
g <- ggplot(data=df, aes(x=Petal.Length, y=Petal.Width))
g <- g + geom_point(aes(color=Species))
g +
theme_dcc() +
scale_fill_dcc()
}
\keyword{scale_fill_dcc}
