% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dcc.R
\name{theme_dcc}
\alias{theme_dcc}
\title{Add DCC theme to ggplot chart}
\usage{
theme_dcc(base_size = 11, base_line_size = base_size/22,
  base_rect_size = base_size/22)
}
\description{
This function allows you to add the DCC theme to your ggplotgraphics.
}
\examples{
df <- iris
g <- ggplot(data=df, aes(x=Petal.Length, y=Petal.Width))
g <- g + geom_point(aes(color=Species))
g + theme_dcc()
}
\keyword{theme_dcc}
