% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_snare.R
\name{theme_snare}
\alias{theme_snare}
\title{Add the Snare Industries theme to your ggplot chart}
\usage{
theme_snare(base_size = 11, base_line_size = base_size/22,
  base_rect_size = base_size/22)
}
\description{
This function allows you to add the Snare Industries theme to your ggplot graphics.
}
\examples{
library(ggplot2)
df <- iris
g <- ggplot(data=df, aes(x=Petal.Length, y=Petal.Width))
g <- g + geom_point(aes(color=Species))
g + theme_snare()
}
\keyword{theme_snare}
